requirements.txt
fastapi
uvicorn
pymongo 
bcrypt 
python-dotenv
python-jose[cryptography]
pydantic[email]
